requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
docx2txt
streamlit-lottie
together